src/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
click==7.1.2
Django==3.1.5
factory-boy==3.2.0
Faker==5.6.5
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.5
regex==2020.11.13
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
wrapt==1.12.1
